minor changes & final version

+++ LibraryApplication/src/main/java/com/example/LibraryApplication/dto/BorrowRequest.java
@@ -1,15 +1,12 @@
 package com.example.LibraryApplication.dto;
 
-import com.example.LibraryApplication.entities.Book;
-import com.example.LibraryApplication.entities.Member;
-
 import lombok.Data;
 import java.util.Date;
 
 @Data
 
 public class BorrowRequest {
-    private int id;
+    private long id;
     private int bookId;
     private int memberId;
     private Date borrowDate;
